Coding style improvements
Primarily type hinting and more consistent use of data types. No
change in functionality.

+++ src/function.py
@@ -5,7 +5,7 @@ from typing import Optional, List
 
 class Function:
     key: str
-    weights: List[float | int]
+    weights: List[int | float]
     is_present: bool
     index: Optional[int]
     input_id: Optional[str]
